Break out  to separate file as fingerprint.rs is big
